[WIP] saca duty cycle

+++ Core/Inc/BCU/Communication/SPI.hpp
@@ -50,13 +50,13 @@ class SPI {
     static bool has_received_unlock_dc_link;
     static bool has_received_commutation_settings;
 
-    double requested_velocity_reference{0.0};
-    double requested_d_current_reference{0.0};
-    double requested_q_current_reference{0.0};
-    double requested_angular_velocity{0.0};
-    double requested_duty_cycle_u{0.0};
-    double requested_duty_cycle_v{0.0};
-    double requested_duty_cycle_w{0.0};
+    float requested_velocity_reference{0.0};
+    float requested_d_current_reference{0.0};
+    float requested_q_current_reference{0.0};
+    float requested_angular_velocity{0.0};
+    float requested_duty_cycle_u{0.0};
+    float requested_duty_cycle_v{0.0};
+    float requested_duty_cycle_w{0.0};
     float forced_dc_link_voltage{0.0};
     uint32_t commutation_frequency_hz{0};
     uint32_t commutation_dead_time_ns{0};
@@ -88,44 +88,44 @@ class SPI {
 
    public:
     SPI(StateMachine::state_id *slave_general_state,
-        StateMachine::state_id *slave_nested_state, double *velocity_reference,
-        double *velocity_error, double *u_current_measurement,
-        double *v_current_measurement, double *w_current_measurement,
-        double *electrical_angle, double *d_current_reference,
-        double *d_current_measurement, double *d_current_error,
-        double *q_current_reference, double *q_current_measurement,
-        double *q_current_error, double *three_phase_unbalance,
-        double *d_target_voltage, double *q_target_voltage,
-        double *u_target_voltage, double *v_target_voltage,
-        double *w_target_voltage, double *u_output_voltage,
-        double *v_output_voltage, double *w_output_voltage,
-        double *u_duty_cycle, double *v_duty_cycle, double *w_duty_cycle,
-        double *angular_velocity, std::array<double *, 3> position,
-        std::array<double *, 3> velocity, std::array<double *, 3> acceleration,
+        StateMachine::state_id *slave_nested_state, float *velocity_reference,
+        float *velocity_error, float *u_current_measurement,
+        float *v_current_measurement, float *w_current_measurement,
+        float *electrical_angle, float *d_current_reference,
+        float *d_current_measurement, float *d_current_error,
+        float *q_current_reference, float *q_current_measurement,
+        float *q_current_error, float *three_phase_unbalance,
+        float *d_target_voltage, float *q_target_voltage,
+        float *u_target_voltage, float *v_target_voltage,
+        float *w_target_voltage, float *u_output_voltage,
+        float *v_output_voltage, float *w_output_voltage, float *u_duty_cycle,
+        float *v_duty_cycle, float *w_duty_cycle, float *angular_velocity,
+        std::array<float *, 3> position, std::array<float *, 3> velocity,
+        std::array<float *, 3> acceleration,
         std::array<Shared::Communication::Direction *, 3> direction,
-        double *average_position, double *max_velocity, bool *is_detecting,
+        float *average_position, float *max_velocity, bool *is_detecting,
         std::array<float *, 4> dc_link_voltage,
         std::array<std::array<float *, 3>, 4> motor_phase_current,
         std::array<PinState *, 4> gate_driver_fault,
         std::array<PinState *, 4> gate_driver_ready);
 
     SPI(Pin &spi_ready_slave_pin, StateMachine::state_id *slave_general_state,
-        StateMachine::state_id *slave_nested_state, double *velocity_reference,
-        double *velocity_error, double *u_current_measurement,
-        double *v_current_measurement, double *w_current_measurement,
-        double *electrical_angle, double *d_current_reference,
-        double *d_current_measurement, double *d_current_error,
-        double *q_current_reference, double *q_current_measurement,
-        double *q_current_error, double *three_phase_unbalance,
-        double *d_target_voltage, double *q_target_voltage,
-        double *u_target_voltage, double *v_target_voltage,
-        double *w_target_voltage, double *u_output_voltage,
-        double *v_output_voltage, double *w_output_voltage,
-        double *u_duty_cycle, double *v_duty_cycle, double *w_duty_cycle,
-        double *angular_velocity, std::array<double *, 3> position,
-        std::array<double *, 3> velocity, std::array<double *, 3> acceleration,
+        StateMachine::state_id *slave_nested_state, float *velocity_reference,
+        float *velocity_error, float *u_current_measurement,
+        float *v_current_measurement, float *w_current_measurement,
+        float *electrical_angle, float *d_current_reference,
+        float *d_current_measurement, float *d_current_error,
+        float *q_current_reference, float *q_current_measurement,
+        float *q_current_error, float *three_phase_unbalance,
+        float *d_target_voltage, float *q_target_voltage,
+        float *u_target_voltage, float *v_target_voltage,
+        float *w_target_voltage, float *u_output_voltage,
+        float *v_output_voltage, float *w_output_voltage, float *u_duty_cycle,
+        float *v_duty_cycle, float *w_duty_cycle, float *angular_velocity,
+        std::array<float *, 3> position, std::array<float *, 3> velocity,
+        std::array<float *, 3> acceleration,
         std::array<Shared::Communication::Direction *, 3> direction,
-        double *average_position, double *max_velocity, bool *is_detecting,
+        float *average_position, float *max_velocity, bool *is_detecting,
         std::array<float *, 4> dc_link_voltage,
         std::array<std::array<float *, 3>, 4> motor_phase_current,
         std::array<PinState *, 4> gate_driver_fault,
